Add initData custom hook
